services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
orjson==3.10.7
